feat: use f32

+++ src/camera.rs
@@ -6,7 +6,7 @@ fn random_in_unit_disk() -> Vec3 {
     let mut rng = rand::thread_rng();
     let unit = Vec3::new(1.0, 1.0, 0.0);
     loop {
-        let p = 2.0 * Vec3::new(rng.gen::<f64>(), rng.gen::<f64>(), 0.0) - unit;
+        let p = 2.0 * Vec3::new(rng.gen::<f32>(), rng.gen::<f32>(), 0.0) - unit;
         if p.dot(&p) < 1.0 {
             return p;
         }
@@ -20,7 +20,7 @@ pub struct Camera {
     vertical: Vec3,
     u: Vec3,
     v: Vec3,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
@@ -28,13 +28,13 @@ impl Camera {
         look_from: Vec3,
         look_at: Vec3,
         view_up: Vec3,
-        vertical_fov: f64,
-        aspect: f64,
-        aperture: f64,
-        focus_dist: f64,
+        vertical_fov: f32,
+        aspect: f32,
+        aperture: f32,
+        focus_dist: f32,
     ) -> Self {
-        let theta = vertical_fov * std::f64::consts::PI / 180.0;
-        let half_height = focus_dist * f64::tan(theta / 2.0);
+        let theta = vertical_fov * std::f32::consts::PI / 180.0;
+        let half_height = focus_dist * f32::tan(theta / 2.0);
         let half_width = aspect * half_height;
         let w = (look_from - look_at).normalize();
         let u = view_up.cross(&w).normalize();
@@ -51,7 +51,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, u: f64, v: f64) -> Ray {
+    pub fn get_ray(&self, u: f32, v: f32) -> Ray {
         let rd = self.lens_radius * random_in_unit_disk();
         let offset = self.u * rd.x + self.v * rd.y;
         Ray::new(
